Two changes, but so interrelated that it is better to commit as one.
1. The 'import_' method returns source (__file__) attribute of the imported
module if available. This is stored as an attribute of test library objects,
and written into syslog as the library is imported.

2. When a test library object is created, if it has either ROBOT_LIBRARY_VERSION
or __version__ attribute, this is stored as an attribute of the test library object.
It is also written in the syslog and e.g. libdoc may use it later.

--HG--
extra : convert_revision : svn%3A79c32731-664e-0410-8185-e51b9e89f9fb/trunk%40650