Implemented `Keyword Should Exist Keyword`
Update issue 896
Status: Done
This is done. Implementation was as simple as I suspected.

In the end I decided to implement this so that you only can give one keyword name plus an optional custom error message. `Variable Should Exists` has the same signature and thought it's better to be consistent. You can always use a for loop if you need to verify that multiple keywords exist.