Add option to `Collections.Lists Should be Equal` to name indices
Update issue 909
Status: Done
Implementation done with tests and docs. Indices can be names either with a list or with a dictionary -- see the doc of the keyword for more details and an example.

The output when indices are named is like this:

Lists are different:
Index 1 (NAME): 2 != 4
Index 2 (Desc): 3 != 8

This is slightly different than requested but ought to provide the same functionality. The reason for the difference is that I implemented this so that you don't need to name all the indices so the error could also be like:

Lists are different:
Index 1 (NAME): 2 != 4
Index 2: 3 != 8