Fixed creating logs/reports when using robotframework.jar. Also some cleanup.
Update issue 916
Yep, generating logs/reports with rf.jar was totally broken. Fixed that and did some other clenup. For the record, the fix was nearly identical to the patch submitted by imran.

I tested this on Linux and it worked fine. Jussi will still test this on Windows and OS X.