Fine tuned enhanced Set Library Search Order
Update issue 935
Now resource/library names in the search order are always case and space insensitive. It used to be case sensitive with libs and both space and case sensitive with resources. The current behavior is documented.

Added explicit tests for setting both library and resource orders at the same time.