Fixed importing Java libraries in packages with robotframework.jar and stand-alone Jython in general.
Update issue 1062
Status: Done
Labels: Type-Task Priority-High
I was able to reproduce the problem and also to fix it. My earlier importing
related fixes and clean-ups had broken the workaround we have for importing
Java classes in packages with stand-alone Jython. For more information
see issue 515 and http://bugs.jython.org/issue1778514.

The reason this bug had slipped through is that we don't have CI environment
for running tests with stand-alone Jython (or robotframework.jar that uses it
internally). Although we ought to have executed these tests at some point
before the final release, it was great that you spotted the bug already at this
point. The fix will be included in the next RF 2.7 preview release (b2 or rc1).