CommandLineMonitor: Notifications on console when top-level keywords end
Update issue 1066
Status: Started
Intial implementation of this feature is now ready. It works so that when
tests are executing, there are dots added after the test name into the
place where the test status is written in the end. A new dot is written
every time when a top-level keyword ends. If there are more dots than
there is room reserved for status, old dots are cleared. Dots are also
always creared when the status is written.

Tasks to do:
- Test how dots work when output is redirected and possibly disable them in that case
- Testing on Windows
- Testing with IronPython
- Investigate could dots be colored based on status
- Review